+++ forui/example/lib/example.dart
@@ -20,43 +20,9 @@ class _ExampleState extends State<Example> {
         mainAxisSize: MainAxisSize.min,
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
+          const SizedBox(height: 100),
+          FProgress(value: 0.9),
           const SizedBox(height: 10),
-          Expanded(
-            child: FTabs(
-              tabs: [
-                FTabEntry(
-                  label: const Text('Account'),
-                  content: FCard(
-                    title: const Text('Account'),
-                    subtitle: const Text('Make changes to your account here. Click save when you are done.'),
-                    child: Column(
-                      children: [
-                        Container(
-                          color: Colors.red,
-                          height: 100,
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-                FTabEntry(
-                  label: const Text('Password'),
-                  content: FCard(
-                    title: const Text('Password'),
-                    subtitle: const Text('Change your password here. After saving, you will be logged out.'),
-                    child: Column(
-                      children: [
-                        Container(
-                          color: Colors.blue,
-                          height: 100,
-                        )
-                      ],
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
         ],
       );
 }