+++ forui/lib/src/theme/theme_data.dart
@@ -40,6 +40,9 @@ final class FThemeData with Diagnosticable {
   /// The header styles.
   final FHeaderStyles headerStyle;
 
+  /// The progress styles.
+  final FProgressStyle progressStyle;
+
   /// The tabs styles.
   final FTabsStyle tabsStyle;
 
@@ -68,6 +71,7 @@ final class FThemeData with Diagnosticable {
     required this.cardStyle,
     required this.dialogStyle,
     required this.headerStyle,
+    required this.progressStyle,
     required this.tabsStyle,
     required this.textFieldStyle,
     required this.scaffoldStyle,
@@ -94,6 +98,7 @@ final class FThemeData with Diagnosticable {
       cardStyle: FCardStyle.inherit(colorScheme: colorScheme, typography: typography, style: style),
       dialogStyle: FDialogStyle.inherit(colorScheme: colorScheme, typography: typography, style: style),
       headerStyle: FHeaderStyles.inherit(colorScheme: colorScheme, typography: typography, style: style),
+      progressStyle: FProgressStyle.inherit(colorScheme: colorScheme, style: style),
       tabsStyle: FTabsStyle.inherit(colorScheme: colorScheme, typography: typography, style: style),
       textFieldStyle: FTextFieldStyle.inherit(colorScheme: colorScheme, typography: typography, style: style),
       scaffoldStyle: FScaffoldStyle.inherit(colorScheme: colorScheme, style: style),
@@ -128,6 +133,7 @@ final class FThemeData with Diagnosticable {
     FCardStyle? cardStyle,
     FDialogStyle? dialogStyle,
     FHeaderStyles? headerStyle,
+    FProgressStyle? progressStyle,
     FTabsStyle? tabsStyle,
     FTextFieldStyle? textFieldStyle,
     FScaffoldStyle? scaffoldStyle,
@@ -143,6 +149,7 @@ final class FThemeData with Diagnosticable {
         cardStyle: cardStyle ?? this.cardStyle,
         dialogStyle: dialogStyle ?? this.dialogStyle,
         headerStyle: headerStyle ?? this.headerStyle,
+        progressStyle: progressStyle ?? this.progressStyle,
         tabsStyle: tabsStyle ?? this.tabsStyle,
         textFieldStyle: textFieldStyle ?? this.textFieldStyle,
         scaffoldStyle: scaffoldStyle ?? this.scaffoldStyle,
@@ -162,6 +169,7 @@ final class FThemeData with Diagnosticable {
       ..add(DiagnosticsProperty('cardStyle', cardStyle, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('dialogStyle', dialogStyle, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('headerStyle', headerStyle, level: DiagnosticLevel.debug))
+      ..add(DiagnosticsProperty('progressStyle', progressStyle))
       ..add(DiagnosticsProperty('tabsStyle', tabsStyle, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('textFieldStyle', textFieldStyle, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('scaffoldStyle', scaffoldStyle, level: DiagnosticLevel.debug))
@@ -182,6 +190,7 @@ final class FThemeData with Diagnosticable {
           cardStyle == other.cardStyle &&
           dialogStyle == other.dialogStyle &&
           headerStyle == other.headerStyle &&
+          progressStyle == other.progressStyle &&
           tabsStyle == other.tabsStyle &&
           textFieldStyle == other.textFieldStyle &&
           scaffoldStyle == other.scaffoldStyle &&
@@ -198,6 +207,7 @@ final class FThemeData with Diagnosticable {
       cardStyle.hashCode ^
       dialogStyle.hashCode ^
       headerStyle.hashCode ^
+      progressStyle.hashCode ^
       tabsStyle.hashCode ^
       textFieldStyle.hashCode ^
       scaffoldStyle.hashCode ^