+++ forui/lib/widgets.dart
@@ -6,6 +6,7 @@ export 'src/widgets/button/button.dart' hide Variant;
 export 'src/widgets/card/card.dart';
 export 'src/widgets/dialog/dialog.dart';
 export 'src/widgets/header/header.dart';
+export 'src/widgets/progress.dart';
 export 'src/widgets/tabs/tabs.dart';
 export 'src/widgets/text_field/text_field.dart';
 export 'src/widgets/scaffold.dart';